Sample files/ssh_keys
sample ansible roles/ common/files/ssh_keys for multi-tier architecture:
Here is an example of a sample ansible role common/files/ssh_keys for a multi-tier architecture:

In this example, the role common/files/ssh_keys is responsible for copying SSH keys to the remote host and adding them to the authorized_keys file for the specified user. The ssh_keys directory should contain the public keys that you want to install on the remote host.