Sample templates/firewall.j2
sample file in directory [ansible roles/ common/templates/firewall.j2] for multi-tier architecture:
Here is an example of the firewall.j2 template that could be used in the common/templates directory for the common role in a multi-tier architecture:

This firewall template generates a simple iptables firewall configuration that accepts incoming packets for established, related connections, ICMP (ping) packets, and packets destined for the loopback interface. It also accepts incoming packets for ports 22 (SSH), 80 (HTTP), and 443 (HTTPS). All other incoming packets are dropped.

You can customize this template to meet the specific security requirements of your multi-tier architecture. For example, you might want to allow additional ports for specific services or restrict access to certain IP addresses.

This template demonstrates how you can use Ansible templates to automate the generation of firewall rules in a multi-tier architecture. The firewall.j2 template can be used in conjunction with the template module in the common/tasks/main.yml file to generate the firewall rules and apply them to the nodes in the common tier.