sample templates/motd.j2
sample file in directory [ansible roles/ common/templates/motd.j2] for multi-tier architecture:
Here is an example of the motd.j2 template that could be used in the common/templates directory for the common role in a multi-tier architecture:

This template generates a message of the day (motd) banner that provides information about the host, including its hostname, operating system and IP address. The tier_name variable can be passed to the template from the Ansible playbook to indicate which tier the node belongs to in the multi-tier architecture.

The motd.j2 template can be used in conjunction with the template module in the common/tasks/main.yml file to generate the motd banner and apply it to the nodes in the common tier. This provides a way to customize the motd banner for each node in the architecture, providing a clear and consistent message to users who log into the nodes.